- outsource all resolution/zoom/viewport-stuff to Qt designer - refine some fontsettings for OSX - make getZoomFactor an integer to have rounded zoomvalues in Gui - redesign resolution/zoom/viewport-stuff as proposed by Abel and Doug

+++ qtgui/renderview.cpp
@@ -108,7 +108,7 @@ void RenderView::resizeEvent(QResizeEvent *event) {
 	emit viewChanged ();
 }
 
-float RenderView::getZoomFactor () {
+int RenderView::getZoomFactor () {
 	return zoomfactor;
 }
